+++ src/pharmpy/modeling/eta_additions.py
@@ -13,11 +13,12 @@ from sympy import Symbol as S
 from sympy.stats.crv_types import NormalDistribution
 from sympy.stats.joint_rv_types import MultivariateNormalDistribution
 
+from pharmpy.expressions import sympify
 from pharmpy.modeling.expressions import create_symbol, get_pk_parameters, has_random_effect
 from pharmpy.modeling.help_functions import _format_input_list, _format_options, _get_etas
 from pharmpy.parameters import Parameter, Parameters
 from pharmpy.random_variables import RandomVariable
-from pharmpy.statements import Assignment, sympify
+from pharmpy.statements import Assignment
 
 
 def add_iiv(
